apply changes for #527

+++ packages/grz-db/src/grz_db/migrations/versions/834bd50b8734_modify_submissions_with_new_column_size_.py
@@ -11,7 +11,6 @@ from collections.abc import Sequence
 import sqlalchemy as sa
 from alembic import op
 
-
 # revision identifiers, used by Alembic.
 revision: str = '834bd50b8734'
 down_revision: str | Sequence[str] | None = '8aa6fc8d118a'
@@ -21,7 +20,7 @@ depends_on: str | Sequence[str] | None = None
 
 def upgrade() -> None:
     """Upgrade schema."""
-    op.add_column("submissions", sa.Column("submission_size", sa.Numeric(10, 2), nullable=True))
+    op.add_column("submissions", sa.Column("submission_size", sa.Integer(), nullable=True))
     op.add_column("submissions", sa.Column("submission_metadata", sa.Text(), nullable=True))
